requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-huggingface>=0.0.3
faiss-cpu==1.7.4
tiktoken>=0.7.0
python-dotenv==1.0.1
sentence-transformers>=2.7.0